fix: remove unintended trigger changes
