+++ src/mock.tsx
@@ -1,1 +1,36 @@
-export { MockTrigger as default, UniqueProvider } from './index';
+import * as React from 'react';
+import { generateTrigger, UniqueProvider } from './index';
+
+interface MockPortalProps {
+  open?: boolean;
+  autoDestroy?: boolean;
+  children: React.ReactElement<any>;
+  getContainer?: () => HTMLElement;
+}
+
+const MockPortal: React.FC<MockPortalProps> = ({
+  open,
+  autoDestroy,
+  children,
+  getContainer,
+}) => {
+  const [visible, setVisible] = React.useState(open);
+
+  React.useEffect(() => {
+    getContainer?.();
+  });
+
+  React.useEffect(() => {
+    if (open) {
+      setVisible(true);
+    } else if (autoDestroy) {
+      setVisible(false);
+    }
+  }, [open, autoDestroy]);
+
+  return visible ? children : null;
+};
+
+export default generateTrigger(MockPortal);
+
+export { UniqueProvider };
